Ping-pong vibro(multi fix)

+++ core/src/main/java/com/logorigm/Constants.java
@@ -3,6 +3,6 @@ package com.logorigm;
 public class Constants {
     public static final Integer SCR_WIDTH = 640;
     public static final Integer SCR_HEIGHT = 480;
-    public static final Integer power = 5;
+    public static final float power = 5f;
     public static final String hit = "Boom!";
 }
